refactor(Authoring): Create CreateNodeBetweenService
